[feature/#13-router] refactoring move Router to folder routing
