+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html dir="ltr" lang="en" xmlns="http://www.w3.org/1999/html"><head> <meta charset="UTF-8"> <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"> <meta http-equiv="X-UA-Compatible" content="ie=edge"> <title>Stencil Component Starter</title> <script type="module" src="./build/honey-news.esm.js"></script> <script nomodule="" src="./build/honey-news.js"></script> <link rel="stylesheet" href="./build/p-9f557766.css"> <!--  <base href="https://huluvu424242.github.io/honey-news/">--> </head> <body id="seite"> <honey-news></honey-news></body></html>
+<!doctype html><html dir="ltr" lang="en" xmlns="http://www.w3.org/1999/html"><head> <meta charset="UTF-8"> <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"> <meta http-equiv="X-UA-Compatible" content="ie=edge"> <title>Stencil Component Starter</title> <script type="module" src="./build/honey-news.esm.js"></script> <script nomodule="" src="./build/honey-news.js"></script> <link rel="stylesheet" href="./build/p-14d21f54.css"> <!--  <base href="https://huluvu424242.github.io/honey-news/">--> </head> <body id="seite"> <honey-news></honey-news></body></html>
